Nois/backend/requirements.txt
Django==5.1.1
django-allauth==64.2.1                # You can uncomment this if you plan to use it
django-cors-headers==4.6.0
django-extensions==3.2.3
django-password-validators==1.7.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
geopy==2.4.1
psycopg2-binary==2.9.9                 # PostgreSQL driver
requests==2.32.3                        # Keep this if your app uses it
pytest==8.3.3                          # Testing libraries
pytest-django==4.9.0
Faker==30.6.0
factory_boy==3.3.1
coverage==7.6.3
python-dotenv==1.0.1
phonenumbers
sib-api-v3-sdk
cryptography
django-debug-toolbar
unidecode